requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-dataclass==8.5.14
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
prisma==0.9.0
py-cord==2.4.1
pycodestyle==2.10.0
pydantic==1.10.9
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.8
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
